// src/domain/characterBase/types/CharacterBaseInformation.ts
type CharacterBaseInformation = {
  name: string;
  class: string;
  level: string;
  sizeCategory: string;
  size: string;
  age: string;
  sex: string;
  playerName: string;
  race: string;
  alignment: string;
  divinity: string;
  weight: string;
  eyes: string;
  hair: string;
  skin: string;
};

export { CharacterBaseInformation };